requirements.txt
Flask==3.1.1
Flask-SQLAlchemy==3.1.1
Flask-Marshmallow==1.3.0
marshmallow==4.0.0
marshmallow-sqlalchemy==1.4.2

Flask-Caching==2.3.1
cachelib==0.13.0

Flask-Limiter==4.0.0
limits==5.6.0

python-dotenv==1.0.1
gunicorn==22.0.0
psycopg2-binary==2.9.9

PyJWT==2.10.1
python-jose==3.3.0

SQLAlchemy==2.0.41
Werkzeug==3.1.3
Jinja2==3.1.6
click==8.2.2
itsdangerous==2.2.0
